Notifications Banner: Refactor & Fix stale height
Split notification file into two modules separating banner from the
normal notification rendering.

Notifications banner used to have one ID which will be persisted to over
multiple notifications making the notification sharing the max height
making new notifications take much more height than needed if we have a
large notification previously.
